Tidy EcdsaWorker type.
So it is no longer declared in the worker which means that we are no
longer statically importing the worker file which is also run
dynamically via `new Worker`.

Use top-level await to initialize the WASM and update typescript
configuration.
